Move functions to future lib.hp
