+++ futurelib.php
@@ -1,0 +1,93 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Backported functions that in a future exists.
+ *
+ * @package    local_mobile
+ * @copyright  2014 Juan Leyva <juan@moodle.com>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+defined('MOODLE_INTERNAL') || die;
+
+require_once($CFG->dirroot . '/mod/chat/lib.php');
+require_once($CFG->dirroot . '/mod/choice/lib.php');
+
+if (!function_exists('chat_get_latest_messages')) {
+
+    /**
+     * Return a list of the latest messages in the given chat session.
+     *
+     * @param  stdClass $chatuser     chat user session data
+     * @param  int      $chatlasttime last time messages were retrieved
+     * @return array    list of messages
+     * @since  Moodle 3.0
+     */
+    function chat_get_latest_messages($chatuser, $chatlasttime) {
+        global $DB;
+
+        $params = array('groupid' => $chatuser->groupid, 'chatid' => $chatuser->chatid, 'lasttime' => $chatlasttime);
+
+        $groupselect = $chatuser->groupid ? " AND (groupid=" . $chatuser->groupid . " OR groupid=0) " : "";
+
+        return $DB->get_records_select('chat_messages_current', 'chatid = :chatid AND timestamp > :lasttime ' . $groupselect,
+                                        $params, 'timestamp ASC');
+    }
+}
+
+if (!function_exists('choice_get_my_response')) {
+    /**
+     * Return my responses on a specific choice.
+     * @param object $choice
+     * @return array
+     */
+    function choice_get_my_response($choice) {
+        global $DB, $USER;
+        return $DB->get_records('choice_answers', array('choiceid' => $choice->id, 'userid' => $USER->id));
+    }
+}
+
+if (!function_exists('choice_can_view_results')) {
+    /**
+     * Return true if we are allowd to see choice results as student
+     * @param object $choice Choice
+     * @param rows|null $current my choice responses
+     * @param bool|null $choiceopen choice open
+     * @return bool True if we can see results, false if not.
+     */
+    function choice_can_view_results($choice, $current = null, $choiceopen = null) {
+
+        if (is_null($choiceopen)) {
+            $timenow = time();
+            if ($choice->timeclose != 0 && $timenow > $choice->timeclose) {
+                $choiceopen = false;
+            } else {
+                $choiceopen = true;
+            }
+        }
+        if (is_null($current)) {
+            $current = choice_get_my_response($choice);
+        }
+
+        if ($choice->showresults == CHOICE_SHOWRESULTS_ALWAYS or
+           ($choice->showresults == CHOICE_SHOWRESULTS_AFTER_ANSWER and !empty($current)) or
+           ($choice->showresults == CHOICE_SHOWRESULTS_AFTER_CLOSE and !$choiceopen)) {
+            return true;
+        }
+        return false;
+    }
+}
